refactor(ui): extract combobox options init. logic into custom hook
