+++ packages/ui-components/src/components/ComboBox/hooks/index.ts
@@ -1,0 +1,5 @@
+export {
+  default as useComboboxOptions,
+  type OptionValuesAndLabelsKey,
+  type OptionValuesAndLabelsValue,
+} from "./useComboBoxOptions"
